refactor(workflow): split phase helpers by concern (§13.2)
`server/workflow/phases/helpers.ts` was 2,337 lines mixing five unrelated
concerns. Every phase module imports from it, so any edit recompiled the lot
and merge conflicts concentrated there. It is now a 16-line barrel of
`export * from`, and no import site changed:

- `logEmission.ts` (247) — tool-log limits, `emitPhaseLog`, `emitDebugLog`,
  `emitModelSystemLog` and the formatting behind them.
- `openCodeStream.ts` (1,199) — stream state, tool-state formatting,
  `emitOpenCodeStreamEvent`, `emitOpenCodeSessionLogs`, question payloads.
- `todoSummary.ts` (33) — `formatTodoTransitionSummary`, the todo-transition
  concern §13.2 names.
- `phaseRuntimeSettings.ts` (290) — the per-ticket settings resolvers and the
  council stage/status mapping.
- `councilDrafts.ts` — draft round summaries, quorum diagnostics, and the
  draft/vote artifact writers.
- `pipelineRecovery.ts` (214) — `tryRecoverPhaseIntermediate` and the
  persisted-draft recovery beneath it.
- `ticketDirContext.ts` (43) — `loadTicketDirContext` and
  `buildStructuredMetadata`, the two helpers three of the above needed. They
  live in their own module rather than staying in the barrel, because a sibling
  importing from the barrel that re-exports it is a cycle.

Departure from the plan: §13.2 asks for `todoSummary.ts` and
`pipelineRecovery.ts` only, which moves ~250 of 2,337 lines and leaves the
recompile-and-conflict problem it describes essentially untouched. Agreed with
the owner to split along the file's real seams instead.

Two consequences worth stating:

- A handful of helpers that were module-private (`stringifyToolDetail`,
  `formatDuration`, `buildBeadLogFields`, `STREAMING_LOG_MIN_INTERVAL_MS`,
  `formatTodoTransitionSummary` and two more) are now exported so their new
  siblings can reach them, which widens the barrel's surface slightly. Nothing
  outside `server/workflow/phases/` imports them.
- `export { resolveAiQuestionSettings }` was carried into all seven modules by
  the split and is now declared once, in `phaseRuntimeSettings.ts`.

Also fixes a source-text gate this stage broke: PR-11's
`protocolTags.test.ts` reads `ArtifactContentViewer.tsx` to prove the client
builds the execution-setup envelope from the constant, and §13.1 moved that
code to `artifactViewers/executionSetupViews.tsx`. The gate now reads the new
path and asserts the file was non-empty first, so the next move fails loudly
rather than passing on an empty read.

Impact: 934 workflow and phase tests pass, 2,140 across the server.
